refactored out subject graph component
a good step to #5 but more refactoring still needs to be done.
Refactored out this component first to address react-force-graph issues
with NextJS' dynamic loading. https://github.com/vasturiano/react-force-graph/issues/324#issuecomment-942725048
